+++ packages/@expo/cli/src/export/publicFolder.ts
@@ -31,6 +31,9 @@ export function getUserDefinedFile(projectRoot: string, possiblePaths: string[])
  */
 export async function copyPublicFolderAsync(publicFolder: string, outputFolder: string) {
   if (fs.existsSync(publicFolder)) {
-    await copyAsync(publicFolder, outputFolder);
+    await fs.promises.mkdir(outputFolder, { recursive: true });
+    await copyAsync(publicFolder, outputFolder, {
+      overwrite: true,
+    });
   }
 }