+++ packages/expo-av/src/Audio/RecordingConstants.ts
@@ -1,35 +1,81 @@
 import type { RecordingOptions } from './Recording.types';
 
-// @docsMissing
+/**
+ * Defines the output format.
+ * @platform android
+ */
 export enum AndroidOutputFormat {
   DEFAULT = 0,
+  /**
+   * 3GPP media file format.
+   */
   THREE_GPP = 1,
+  /**
+   * MPEG4 media file format.
+   */
   MPEG_4 = 2,
+  /**
+   * AMR NB file format.
+   */
   AMR_NB = 3,
+  /**
+   * AMR WB file format.
+   */
   AMR_WB = 4,
+  // @docsMissing
   AAC_ADIF = 5,
+  /**
+   * AAC ADTS file format.
+   */
   AAC_ADTS = 6,
+  // @docsMissing
   RTP_AVP = 7,
+  /**
+   * H.264/AAC data encapsulated in MPEG2/TS.
+   */
   MPEG2TS = 8,
+  /**
+   * VP8/VORBIS data in a WEBM container.
+   */
   WEBM = 9,
 }
 
-// @docsMissing
+/**
+ * Defines the audio encoding.
+ * @platform android
+ */
 export enum AndroidAudioEncoder {
   DEFAULT = 0,
+  /**
+   * AMR (Narrowband) audio codec.
+   */
   AMR_NB = 1,
+  /**
+   * AMR (Wideband) audio codec.
+   */
   AMR_WB = 2,
+  /**
+   * AAC Low Complexity (AAC-LC) audio codec.
+   */
   AAC = 3,
+  /**
+   * High Efficiency AAC (HE-AAC) audio codec.
+   */
   HE_AAC = 4,
+  /**
+   * Enhanced Low Delay AAC (AAC-ELD) audio codec.
+   */
   AAC_ELD = 5,
 }
 
 // @docsMissing
 /**
- * > **Note** Not all of the iOS formats included in this list of constants are currently supported by iOS,
+ * > **Note:** Not all of the iOS formats included in this list of constants are currently supported by iOS,
  * > in spite of appearing in the Apple source code. For an accurate list of formats supported by iOS, see
  * > [Core Audio Codecs](https://developer.apple.com/library/content/documentation/MusicAudio/Conceptual/CoreAudioOverview/CoreAudioEssentials/CoreAudioEssentials.html)
  * > and [iPhone Audio File Formats](https://developer.apple.com/library/content/documentation/MusicAudio/Conceptual/CoreAudioOverview/CoreAudioEssentials/CoreAudioEssentials.html).
+ *
+ * @platform ios
  */
 export enum IOSOutputFormat {
   LINEARPCM = 'lpcm',
@@ -69,6 +115,9 @@ export enum IOSOutputFormat {
 }
 
 // @docsMissing
+/**
+ * @platform ios
+ */
 export enum IOSAudioQuality {
   MIN = 0,
   LOW = 0x20,
@@ -78,6 +127,9 @@ export enum IOSAudioQuality {
 }
 
 // @docsMissing
+/**
+ * @platform ios
+ */
 export enum IOSBitRateStrategy {
   CONSTANT = 0,
   LONG_TERM_AVERAGE = 1,